Tags are no longer required to provide a renderer {M,F,A}
Some tags, such as 'comment', don't need a render function. Up till
now they've been providing a render function that returns "". With
this commit, the parse function for a tag now has the choice of
returning, {{Mod, Fun, Args}, #ps{}} or just #ps{} if it does not
need to render anything at runtime.