Replace template unique id with a content-derived MD5 checksum
Having a unique id for each template is not useful when it comes to
detecting recursive template inclusion. The problem is that many
files that are included at runtime are compiled on the fly. If the
same source file was to be given a unique id each time it was
compiled, this id would be ineffective in detecting recursive loops.

An MD5 checksum which is based on a normalized version of the
source code will not change. We will use this for loop detection
instead of a unique ID. Note that an MD5 based on the compiled
template could change because the template generates unique ids
for the benefit of certain tags (such as the 'cycle' tag). For
this reason we base the MD5 on the source text.